Move code to init, add __main__
this enables you to do `python -m vsgenstubs4`
