+++ vsgenstubs4/__main__.py
@@ -1,0 +1,3 @@
+from .init import sys, main
+
+main(sys.argv[1:])
